simeple mouse look for raycasting_1 demo

+++ demos/raycasting_1/src/render/sw/SwRenderer.cpp
@@ -147,32 +147,32 @@ void SwRenderer::DrawPlayerView(const Level& level, const Player& player)
     {
         // Current column position relative to the center of the screen.
         // Left edge is -1, right edge is 1, and center is 0.
-        const double cam_x = 2. * x / mResX - 1;
+        const float cam_x = 2.f * x / mResX - 1;
 
         // Starting direction and  position of the current ray to be cast.
-        const double ray_dir_x = player.mDirX + player.mPlaneX * cam_x;
-        const double ray_dir_y = player.mDirY + player.mPlaneY * cam_x;
-        const double ray_pos_x = player.mPosX;
-        const double ray_pos_y = player.mPosY;
+        const float ray_dir_x = player.mDirX + player.mPlaneX * cam_x;
+        const float ray_dir_y = player.mDirY + player.mPlaneY * cam_x;
+        const float ray_pos_x = player.mPosX;
+        const float ray_pos_y = player.mPosY;
 
         // The direction to step in X or Y-direction (either +1 or -1).
         const int step_x = (ray_dir_x >= 0) ? 1 : -1;
         const int step_y = (ray_dir_y >= 0) ? 1 : -1;
 
         // Length of ray from one X and Y-side to next X and Y-side.
-        const double delta_dist_x = std::sqrt(1 + std::pow(ray_dir_y, 2) / std::pow(ray_dir_x, 2));
-        const double delta_dist_y = std::sqrt(1 + std::pow(ray_dir_x, 2) / std::pow(ray_dir_y, 2));
+        const float delta_dist_x = std::sqrt(1 + std::pow(ray_dir_y, 2) / std::pow(ray_dir_x, 2));
+        const float delta_dist_y = std::sqrt(1 + std::pow(ray_dir_x, 2) / std::pow(ray_dir_y, 2));
 
         // The player's current grid position inside the level.
         int map_x = ray_pos_x;
         int map_y = ray_pos_y;
 
         // Length of ray from its current position to next X- and Y-side.
-        double side_dist_x = (step_x == 1) ?
-                                ((map_x + 1.0 - ray_pos_x) * delta_dist_x) :
+        float side_dist_x = (step_x == 1) ?
+                                ((map_x + 1.0f - ray_pos_x) * delta_dist_x) :
                                 ((ray_pos_x - map_x) * delta_dist_x);
-        double side_dist_y = (step_y == 1) ?
-                                ((map_y + 1.0 - ray_pos_y) * delta_dist_y) :
+        float side_dist_y = (step_y == 1) ?
+                                ((map_y + 1.0f - ray_pos_y) * delta_dist_y) :
                                 ((ray_pos_y - map_y) * delta_dist_y);
 
         // Y walls (EW) will be drawn darker.
@@ -201,12 +201,12 @@ void SwRenderer::DrawPlayerView(const Level& level, const Player& player)
 
         // Calculate the perpendicular distance projected on camera direction.
         // Oblique distance would give fisheye effect.
-        const double perp_wall_dist = y_side_hit ?
+        const float perp_wall_dist = y_side_hit ?
                std::abs((map_y - ray_pos_y + (1 - step_y) / 2) / ray_dir_y) :
                std::abs((map_x - ray_pos_x + (1 - step_x) / 2) / ray_dir_x);
 
         // Calculate the height of the vertical line to draw on screen.
-        const double line_height_d = std::abs(mResY / perp_wall_dist);
+        const float line_height_d = std::abs(mResY / perp_wall_dist);
         const int line_height = line_height_d;
 
         // FIXME: line_height might be negative when the player is near a wall
